requirement.txt
python-telegram-bot==20.4
rivescript==1.15.0
